Fix delay on boss split
Since detecting that the boss has died, instead of the player, can be
somewhat finicky, there was a 5 frame delay to ensure that the boss had
actually died.

Keep track of the player's health to detect whether the boss has died or
whether the battle is restarting because the player died.