Make splits consistent in every level (and boss)
Except for Level 4, the boss is always in the final part of the stage.
However, Level 4 is divided in two sub-stages and the boss is actually at
the end of the first stage. Since the cutscene before and after that fight
has two bosses (used as actors/NPCs in the cutscene), checking whether the
boss pointer is NULL or not was quite finicky. So, to continue splitting
after the boss, this split was moved to right after the last hit is dealt
to the boss.

This required tracking the boss health, and thus enable optionally
splitting after mid-bosses and Yuyuko's first phase.

In every other case, the script simply keeps track of which is the next
expected level and split as soon as the game reaches that.