Move the Robot split to the start of the next scene
The Robot split is somewhat finicky... Both the Factory and the River
levels are seen as "Level 4" by the game.

So instead of splitting when the level changed or when the boss is
defeated, temporarily track that the Robot was defeated sometime ago and
then split when the next scene starts loading.